Abstraction & Reading from config & Scaling to fit the window & Resize events

+++ src/Models/Point.cs
@@ -6,8 +6,8 @@ namespace VectorViewer.Models
     [JsonConverter(typeof(PointConverter))]
     public struct Point
     {
-        public long X { get; set; }
-        public long Y { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
 
         public System.Windows.Point ToWindowsPoint() => new() { X = X, Y = Y };
     }
